+++ changes.md
@@ -1,5 +1,11 @@
 # Morbid
 
+## Release v1.4.0
+LTS: 24/09/2025
+
+ - Temporarily disabled insert users without id in repo
+ - Using UID from Firebase
+
 ## Release v1.3.0
 LTS: 22/09/2025